requirements.txt
botorch>=0.7.3
evotorch>=0.3.0
gpytorch>=1.9.0
matplotlib>=3.6.2
numpy>=1.23.5
gym==0.21.0
pyglet==1.5.27
box2d-py==2.3.8